Finished functionality of class (Need to add in-line comments)

+++ src/main/java/cs1302/api/Recipe.java
@@ -1,10 +1,14 @@
 package cs1302.api;
 
+/**
+ * Class that represents the items in a recipe object
+ * from query.
+ */
 public class Recipe {
     String label;
     String image;
     String url;
-    int yield;
+    double yield;
     String[] healthLabels;
     String[] ingredientLines;
 } // Recipe
